Move Policies to separate file
